+++ src/server.rs
@@ -295,6 +295,7 @@ where
                     let mut svc = HttpService::build()
                         .keep_alive(c.keep_alive)
                         .client_timeout(c.client_timeout)
+                        .client_disconnect(c.client_shutdown)
                         .local_addr(addr);
 
                     if let Some(handler) = on_connect_fn.clone() {
@@ -352,7 +353,8 @@ where
                     let svc = HttpService::build()
                         .keep_alive(c.keep_alive)
                         .client_timeout(c.client_timeout)
-                        .client_disconnect(c.client_shutdown);
+                        .client_disconnect(c.client_shutdown)
+                        .local_addr(addr);
 
                     let svc = if let Some(handler) = on_connect_fn.clone() {
                         svc.on_connect_ext(move |io: &_, ext: _| {
@@ -523,10 +525,11 @@ where
             addr: socket_addr,
         });
 
-        let addr = format!("actix-web-service-{:?}", lst.local_addr()?);
+        let addr = lst.local_addr()?;
+        let name = format!("actix-web-service-{:?}", addr);
         let on_connect_fn = self.on_connect_fn.clone();
 
-        self.builder = self.builder.listen_uds(addr, lst, move || {
+        self.builder = self.builder.listen_uds(name, lst, move || {
             let c = cfg.lock().unwrap();
             let config = AppConfig::new(
                 false,
@@ -537,7 +540,8 @@ where
             fn_service(|io: UnixStream| async { Ok((io, Protocol::Http1, None)) }).and_then({
                 let mut svc = HttpService::build()
                     .keep_alive(c.keep_alive)
-                    .client_timeout(c.client_timeout);
+                    .client_timeout(c.client_timeout)
+                    .client_disconnect(c.client_shutdown);
 
                 if let Some(handler) = on_connect_fn.clone() {
                     svc = svc
@@ -554,8 +558,8 @@ where
         Ok(self)
     }
 
-    #[cfg(unix)]
     /// Start listening for incoming unix domain connections.
+    #[cfg(unix)]
     pub fn bind_uds<A>(mut self, addr: A) -> io::Result<Self>
     where
         A: AsRef<std::path::Path>,
@@ -568,6 +572,7 @@ where
         let factory = self.factory.clone();
         let socket_addr =
             net::SocketAddr::new(net::IpAddr::V4(net::Ipv4Addr::new(127, 0, 0, 1)), 8080);
+
         self.sockets.push(Socket {
             scheme: "http",
             addr: socket_addr,
@@ -592,6 +597,7 @@ where
                     HttpService::build()
                         .keep_alive(c.keep_alive)
                         .client_timeout(c.client_timeout)
+                        .client_disconnect(c.client_shutdown)
                         .finish(map_config(fac, move |_| config.clone())),
                 )
             },